project/requirements.txt
aniso8601==10.0.1
ansi2html==1.9.2
arrow==1.3.0
attrs==23.2.0
Babel==2.10.3
blinker==1.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
enum-compat==0.0.3
filelock==3.18.0
Flask==3.1.0
Flask-RESTful==0.3.10
fonttools==4.57.0
fsspec==2025.3.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kagglehub==0.3.12
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
nvgpu==0.10.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.3.0
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.7.0
torch-model-archiver==0.12.0
torchserve==0.12.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
